# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Ecommerce API
  version: 0.0.1-SNAPSHOT
  description: |
    OpenAPI specification for the ecommerce backend.
    The contract is derived from the Spring controllers in this repository.
  contact:
    name: Abhishek
    email: aabhiojha08@outlook.com
    url: https://abhishekojha.com.np
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/images:
    get:
      tags: [Images]
      summary: List all uploaded images
      operationId: listImages
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of images.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ImageDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/images/upload:
    post:
      tags: [Images]
      summary: Upload an image for a product
      operationId: uploadProductImage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ImageUploadRequest'
            encoding:
              file:
                contentType: application/octet-stream
      responses:
        '200':
          description: Image uploaded successfully.
          content:
            text/plain:
              schema:
                type: string
              example: Uploaded successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/images/upload-multiple:
    post:
      tags: [Images]
      summary: Upload multiple images for a product
      operationId: uploadProductImages
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ImageUploadBatchRequest'
            encoding:
              files:
                contentType: application/octet-stream
      responses:
        '200':
          description: Images uploaded successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ImageDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Validation or business rule failure.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication is required or the token is invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: The authenticated user does not have permission to access the resource.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        error:
          type: string
        message:
          type: string
        path:
          type: string
      required:
        - status
        - error
        - message
        - path
...
    ImageDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        fileName:
          type: string
        fileType:
          type: string
        downloadUrl:
          type: string
        productId:
          type: integer
          format: int64
      required:
        - id
        - fileName
        - fileType
        - downloadUrl
        - productId
...
    ImageUploadRequest:
      type: object
      properties:
        file:
          type: string
          format: binary
        productId:
          type: integer
          format: int64
      required:
        - file
        - productId
    ImageUploadBatchRequest:
      type: object
      properties:
        files:
          type: array
          items:
            type: string
            format: binary
        productId:
          type: integer
          format: int64
      required:
        - files
        - productId
...